Modifications

+++ src/main/controller/MockProduct.java
@@ -1,5 +1,5 @@
 package main.controller;
 
 public interface MockProduct {
-    public float auxiliarProduct(float x, float y); // Funció sense definir que modifiquem el comportament al test
+    public double auxiliarProduct(double x, double y); // Funció sense definir que modifiquem el comportament al test
 }
